+++ leda/components/VStepper/constants.ts
@@ -1,0 +1,4 @@
+export enum Display {
+  Block = 'block',
+  None = 'none'
+}